Block creating plans on the plans table unless we have sufficient providers
