+++ src/app/Plans/components/CreatePlanButton.tsx
@@ -1,0 +1,38 @@
+import * as React from 'react';
+import { Button, ButtonProps } from '@patternfly/react-core';
+import spacing from '@patternfly/react-styles/css/utilities/Spacing/spacing';
+import AddTooltip from '@app/common/components/AddTooltip';
+import { useHasSufficientProvidersQuery } from '@app/queries';
+import { useHistory } from 'react-router-dom';
+
+interface ICreatePlanButtonProps {
+  variant?: ButtonProps['variant'];
+  label?: string;
+}
+
+const CreatePlanButton: React.FunctionComponent<ICreatePlanButtonProps> = ({
+  variant = 'primary',
+  label = 'Create migration plan',
+}: ICreatePlanButtonProps) => {
+  const sufficientProvidersQuery = useHasSufficientProvidersQuery();
+  const { hasSufficientProviders } = sufficientProvidersQuery;
+  const history = useHistory();
+  return (
+    <AddTooltip
+      isTooltipEnabled={!hasSufficientProviders}
+      content="You must add at least one VMware provider and one OpenShift Virtualization provider in order to create a migration plan."
+    >
+      <div className={`${spacing.mtMd}`}>
+        <Button
+          onClick={() => history.push('/plans/create')}
+          isDisabled={!hasSufficientProviders}
+          variant={variant}
+        >
+          {label}
+        </Button>
+      </div>
+    </AddTooltip>
+  );
+};
+
+export default CreatePlanButton;
